requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
